Added more details to the publishing section
I added 2 sub-sections to the Publishing section that give more detail regarding publishing an extension for the first time as well as how to publish changes to your extension. It goes into detail about the version number and about how a publish will be rejected if the version number hasn't been changed, etc.